refactored instruction handlers to solve cyclic import via dependency injection
