+++ prodigy/analysis/instructionhandler/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from probably.pgcl import Instr
+
+
+def _assume(instruction: Instr, instr_type, clsname: str):
+    """ Checks that the instruction is given as the right type."""
+    assert isinstance(instruction, instr_type), f"The Instruction handled by a {clsname} must be of type" \
+                                                f" {instr_type} got {type(instruction)}"
