+++ prodigy/analysis/instructionhandler/query_handler.py
@@ -1,0 +1,138 @@
+from __future__ import annotations
+
+import logging
+from typing import get_args, Callable, Union, Sequence
+
+from probably.pgcl import Instr, Query, ExpectationInstr, VarExpr, NatLitExpr, RealLitExpr, BinopExpr, Binop, UnopExpr, \
+    Unop, ProbabilityQueryInstr, PlotInstr, PrintInstr, OptimizationQuery, OptimizationType, Expr
+
+from prodigy.analysis.config import ForwardAnalysisConfig
+from prodigy.analysis.instructionhandler import _assume
+from prodigy.analysis.instructionhandler.instruction_handler import InstructionHandler
+from prodigy.analysis.instructionhandler.program_info import ProgramInfo
+from prodigy.distribution import Distribution, MarginalType
+from prodigy.util.logger import log_setup
+from prodigy.util.plotter import Plotter
+
+logger = log_setup(__name__, logging.DEBUG)
+
+
+class QueryHandler(InstructionHandler):
+    @staticmethod
+    def compute(
+            instruction: Instr, prog_info: ProgramInfo,
+            distribution: Distribution, error_prob: Distribution,
+            config: ForwardAnalysisConfig,
+            analyzer: Callable[
+                [Union[Instr, Sequence[Instr]], ProgramInfo, Distribution, Distribution, ForwardAnalysisConfig],
+                tuple[Distribution, Distribution]
+            ]
+    ) -> tuple[Distribution, Distribution]:
+        _assume(instruction, get_args(Query), 'QueryHandler')
+
+        # User wants to compute an expected value of an expression
+        if isinstance(instruction, ExpectationInstr):
+            expression = instruction.expr
+            if isinstance(expression, (VarExpr, NatLitExpr, RealLitExpr)):
+                result = distribution.get_expected_value_of(expression)
+            elif isinstance(expression, BinopExpr):
+                if expression.operator in (Binop.PLUS, Binop.MINUS,
+                                           Binop.TIMES):
+                    result = distribution.get_expected_value_of(expression)
+                else:
+                    raise SyntaxError("Expression cannot be a condition.")
+            elif isinstance(expression,
+                            UnopExpr) and expression.operator == Unop.NEG:
+                result = distribution.get_expected_value_of(expression)
+            else:
+                raise SyntaxError("Expression has wrong format.")
+
+            print(f"Expected value: {result}")
+            return distribution, error_prob
+
+        # User wants to compute a marginal, or the probability of a condition.
+        elif isinstance(instruction, ProbabilityQueryInstr):
+            return QueryHandler.__query_probability_of(instruction.expr,
+                                                       distribution), \
+                error_prob
+
+        # User wants to Plot something
+        elif isinstance(instruction, PlotInstr):
+            return QueryHandler.__query_plot(instruction, distribution), \
+                error_prob
+
+        # User wants to print the current distribution.
+        elif isinstance(instruction, PrintInstr):
+            print(distribution)
+            return distribution, error_prob
+
+        elif isinstance(instruction, OptimizationQuery):
+            return QueryHandler.__query_optimization(instruction, distribution,
+                                                     config), \
+                error_prob
+
+        else:
+            raise SyntaxError("This should not happen.")
+
+    @staticmethod
+    def __query_optimization(instr: OptimizationQuery, dist: Distribution,
+                             config: ForwardAnalysisConfig):
+        logger.debug(
+            "Computing the optimal value for parameter %s in order to %s the distribution %s with respect to %s",
+            instr.parameter, 'maximize' if instr.type == OptimizationType.MAXIMIZE else 'minimize', dist, instr.expr)
+        result = config.optimizer.optimize(instr.expr,
+                                           dist,
+                                           instr.parameter,
+                                           method=instr.type)
+        if not result:
+            print("No solutions could be found.")
+        elif len(result) == 1:
+            print(f"The optimal value is at {instr.parameter}={result[0]}.")
+        else:
+            print(
+                f"The maximal values can be achieved choosing the parameter {instr.parameter} in {result}."
+            )
+        return dist
+
+    @staticmethod
+    def __query_plot(instr: PlotInstr, dist: Distribution) -> Distribution:
+        # Gather the variables to plot.
+        variables = [instr.var_1.var]
+        variables += [instr.var_2.var] if instr.var_2 else []
+
+        # User can specify either a probability threshold or a number of terms which are shown in the histogram.
+        if instr.prob:
+            # User chose a probability or oo (meaning, try to compute the whole histogram).
+            if instr.prob.is_infinite():
+                Plotter.plot(dist, *variables, threshold=str(1))
+            else:
+                Plotter.plot(dist, *variables, threshold=str(instr.prob))
+
+        elif instr.term_count:
+            # User has chosen a term limit.
+            Plotter.plot(dist, *variables, threshold=instr.term_count.value)
+            # User did neither specify a term limit nor a threshold probability. We try an iterative approach...
+        else:
+            p = .1
+            inc = .1
+            while True:
+                Plotter.plot(dist, *variables, threshold=str(p))
+                p = p + inc if p + inc < 1 else 1
+                cont = input(f"Continue with p={p}? [Y/n]")
+                if cont.lower() == 'n':
+                    break
+        return dist
+
+    @staticmethod
+    def __query_probability_of(expression: Expr,
+                               dist: Distribution) -> Distribution:
+        # Marginal computation
+        if isinstance(expression, VarExpr):
+            marginal = dist.marginal(expression, method=MarginalType.INCLUDE)
+            print(f"Marginal distribution of {expression}: {marginal}")
+        # Probability of condition computation.
+        else:
+            sat_part = dist.filter(expression)
+            prob = sat_part.get_probability_mass()
+            print(f"Probability of {expression}: {prob}")
+        return dist
