+++ tests/analysis/test_equivalence_check.py
@@ -3,6 +3,7 @@ import sympy
 from probably.pgcl.ast import Program
 from probably.pgcl.compiler import compile_pgcl
 
+from prodigy.analysis.analyzer import compute_semantics
 from prodigy.analysis.config import ForwardAnalysisConfig
 from prodigy.analysis.equivalence.equivalence_check import check_equivalence
 
@@ -36,8 +37,7 @@ def test_equivalence_check(engine):
         } else {skip}
     """)
     assert isinstance(inv, Program)
-    res, subs = check_equivalence(prog, inv,
-                                  ForwardAnalysisConfig(engine=engine))
+    res, subs = check_equivalence(prog, inv, ForwardAnalysisConfig(engine=engine), compute_semantics)
     assert res
     assert subs == []
 
@@ -73,8 +73,7 @@ def test_equivalence_check_parameter(engine):
         } else {skip}
     """)
     assert isinstance(inv, Program)
-    res, subs = check_equivalence(prog, inv,
-                                  ForwardAnalysisConfig(engine=engine))
+    res, subs = check_equivalence(prog, inv, ForwardAnalysisConfig(engine=engine), compute_semantics)
     assert res
     assert len(subs) == 1
     assert sympy.S(subs[0]['p']) == sympy.S('1/2')
